+++ miden-lib/src/accounts/wallets/mod.rs
@@ -15,7 +15,7 @@ use super::{AuthScheme, TransactionKernel};
 // BASIC WALLET
 // ================================================================================================
 
-/// Creates a new account with basic wallet interface and the specified authentication scheme.
+/// Creates a new account with basic wallet interface, the specified authentication scheme and the account storage type.
 /// Basic wallets can be specified to have either mutable or immutable code.
 ///
 /// The basic wallet interface exposes two procedures:
@@ -30,6 +30,7 @@ pub fn create_basic_wallet(
     init_seed: [u8; 32],
     auth_scheme: AuthScheme,
     account_type: AccountType,
+    account_storage_type: AccountStorageType,
 ) -> Result<(Account, Word), AccountError> {
     if matches!(account_type, AccountType::FungibleFaucet | AccountType::NonFungibleFaucet) {
         return Err(AccountError::AccountIdInvalidFieldElement(
@@ -68,7 +69,7 @@ pub fn create_basic_wallet(
     let account_seed = AccountId::get_account_seed(
         init_seed,
         account_type,
-        AccountStorageType::OffChain,
+        account_storage_type,
         account_code.root(),
         account_storage.root(),
     )?;